Break out SSH related into GitSshHelpers
