+++ GitCommands/Git/GitCommandHelpers.cs
@@ -17,53 +17,6 @@ namespace GitCommands
 {
     public static class GitCommandHelpers
     {
-        #region SSH / Plink
-
-        private static readonly ISshPathLocator _sshPathLocatorInstance = new SshPathLocator();
-
-        public static bool UseSsh(string arguments)
-        {
-            var x = !Plink() && DoArgumentsRequireSsh();
-            return x || arguments.Contains("plink");
-
-            bool DoArgumentsRequireSsh()
-            {
-                return (arguments.Contains("@") && arguments.Contains("://")) ||
-                       (arguments.Contains("@") && arguments.Contains(":")) ||
-                       arguments.Contains("ssh://") ||
-                       arguments.Contains("http://") ||
-                       arguments.Contains("git://") ||
-                       arguments.Contains("push") ||
-                       arguments.Contains("remote") ||
-                       arguments.Contains("fetch") ||
-                       arguments.Contains("pull");
-            }
-        }
-
-        /// <summary>Un-sets the git SSH command path.</summary>
-        public static void UnsetSsh()
-        {
-            Environment.SetEnvironmentVariable("GIT_SSH", "", EnvironmentVariableTarget.Process);
-        }
-
-        /// <summary>Sets the git SSH command path.</summary>
-        public static void SetSsh(string path)
-        {
-            if (!string.IsNullOrEmpty(path))
-            {
-                Environment.SetEnvironmentVariable("GIT_SSH", path, EnvironmentVariableTarget.Process);
-            }
-        }
-
-        public static bool Plink()
-        {
-            var sshString = _sshPathLocatorInstance.Find(AppSettings.GitBinDir);
-
-            return sshString.EndsWith("plink.exe", StringComparison.CurrentCultureIgnoreCase);
-        }
-
-        #endregion
-
         public static ArgumentString CherryPickCmd(ObjectId commitId, bool commit, string arguments)
         {
             return new GitArgumentBuilder("cherry-pick")
