+++ GitUI/CommandsDialogs/FormDeleteRemoteBranch.cs
@@ -110,7 +110,7 @@ namespace GitUI.CommandsDialogs
 
         private void EnsurePageant(string remote)
         {
-            if (GitCommandHelpers.Plink())
+            if (GitSshHelpers.Plink())
             {
                 if (!File.Exists(AppSettings.Pageant))
                 {
