+++ GitUI/HelperDialogs/FormRemoteProcess.cs
@@ -67,7 +67,7 @@ Do you want to register the host's fingerprint and restart the process?");
         protected override void BeforeProcessStart()
         {
             _restart = false;
-            Plink = GitCommandHelpers.Plink();
+            Plink = GitSshHelpers.Plink();
             base.BeforeProcessStart();
         }
